eleccion y menu principal, terminados por ahora. Corregir nombre en navbar q sea tiempo real y creacion pagina settings

+++ src/app/colegio/menu-principal/menu-principal.component.ts
@@ -14,7 +14,7 @@ import { Colegio } from 'src/app/shared/interface/user.interface';
 export class MenuPrincipalComponent implements OnInit {
   nombreColegio: string;
   nombreDocumento: string;
-  borroColegio: boolean = true;
+  noBorroColegio: boolean = true;
   // duracionModulo: number;
 
   constructor(
@@ -49,7 +49,7 @@ export class MenuPrincipalComponent implements OnInit {
                 // this.duracionModulo = school.duracionModulo;
               }
               else{
-                this.borroColegio = false;
+                this.noBorroColegio = false;
               }
             })
           )
@@ -69,6 +69,8 @@ export class MenuPrincipalComponent implements OnInit {
   }
 
   async deleteSchool() {
-    this.afs.collection('schools').doc(this.nombreDocumento).delete();
+    if (confirm('¿Estas seguro/a que quieres eliminar este colegio?')) {
+      this.afs.collection('schools').doc(this.nombreDocumento).delete();
+    }
   }
 }
